refactor[#523] : Marking 타입 import 경로 수정
