+++ src/mocks/data/userMarkingList.ts
@@ -1,4 +1,4 @@
-import type { Marking } from "@/entities/marking/api";
+import type { Marking } from "@/entities/marking/types/server";
 
 export const getMockUserMarkingList = ({
   nickname,
